fix: Change add_route view to only allow superuser to add route. Minor code changes to allow image upload

+++ bookings/views.py
@@ -4,6 +4,7 @@ from .models import Route, Trip
 from .forms import RouteForm, TripForm
 from django.http import HttpResponseRedirect
 from django.contrib import messages
+from cloudinary.forms import cl_init_js_callbacks 
 
 
 # Render Home Page
@@ -16,7 +17,7 @@ class RouteList(ListView):
     model = Route
     queryset = Route.objects.filter(status=1).order_by('route_name')
     template_name = 'routes.html'
-    paginate_by = 4
+    paginate_by = 6
 
 
 # Create Trips Views
@@ -40,37 +41,49 @@ class Trips(View):
 
 # Superuser can view all routes on the database from the frontend
 def admin_panel(request):
-    routes = Route.objects.all()
-    trips = Trip.objects.all()
+    if request.user.is_superuser:
+        routes = Route.objects.all()
+        trips = Trip.objects.all()
 
-    context = {
-            'routes': routes,
-            'trips': trips,
-            }
+        context = {
+                'routes': routes,
+                'trips': trips,
+                }
 
-    return render(request, 'admin_panel.html', context=context)
+        return render(request, 'admin_panel.html', context=context)
+
+    else:
+        messages.success(request, (
+            'Access denied. Please sign in as an admin.'))
+        return redirect('account_login')
 
 
 # Superuser can add a route to the database from the frontend
 def add_route(request):
-    submitted = False
+    if request.user.is_superuser:
+        submitted = False
 
-    if request.method == "POST":
-        form = RouteForm(request.POST)
-        if form.is_valid():
-            form.save()
-            return HttpResponseRedirect('add_route?submitted=True')
+        if request.method == "POST":
+            form = RouteForm(request.POST, request.FILES)
+            if form.is_valid():
+                form.save()
+                return HttpResponseRedirect('add_route?submitted=True')
+        else:
+            form = RouteForm
+            if 'submitted' in request.GET:
+                submitted = True
+
+        context = {
+            'form': form,
+            'submitted': submitted
+            }
+
+        return render(request, 'add_route.html', context=context)
+
     else:
-        form = RouteForm
-        if 'submitted' in request.GET:
-            submitted = True
-
-    context = {
-        'form': form,
-        'submitted': submitted
-        }
-
-    return render(request, 'add_route.html', context=context)
+        messages.success(request, (
+            'Access denied. Please sign in as an admin.'))
+        return redirect('account_login')
 
 
 # Superuser can add a trip to the database from the frontend
@@ -151,17 +164,23 @@ def delete_route(request, route_id):
 
 # Superuser can delete a Trip
 def delete_trip(request, trip_id):
-    trip = Trip.objects.get(id=trip_id)
+    if request.user.is_superuser:
+        trip = Trip.objects.get(id=trip_id)
 
-    # Check that user really wants to delete this trip!
-    if request.method == 'POST':
-        trip.delete()
+        # Check that user really wants to delete this trip!
+        if request.method == 'POST':
+            trip.delete()
+            messages.success(request, (
+                'Success! The trip has been deleted from the database'))
+            return redirect('admin-panel')
+
+        context = {
+                'trip': trip
+                }
+
+        return render(request, 'delete_trip.html', context=context)
+    
+    else:
         messages.success(request, (
-            'Success! The trip has been deleted from the database'))
-        return redirect('admin-panel')
-
-    context = {
-            'trip': trip
-            }
-
-    return render(request, 'delete_trip.html', context=context)
+            'You do not have permission to access this page. Please sign in as an admin.'))
+        return redirect('home')
